client phone number

+++ database/migrations/2021_05_20_191802_create_clients_table.php
@@ -18,7 +18,7 @@ class CreateClientsTable extends Migration
             $table->string('name');
             $table->string('last_name');
             $table->string('email')->unique();
-            $table->integer('phone_number');
+            $table->bigInteger('phone_number');
             $table->timestamps();
             $table->unsignedInteger('user_id');
